RenderButton name refactors
Clear distinction between SidebarRenderButton and the main RenderButton
